Move MultiBinder to top level
